LayoutTests/svg/dynamic-updates/SVGFECompositeElement-svgdom-k2-prop-expected.txt

CONSOLE MESSAGE: line 118: TypeError: 'undefined' is not an object (evaluating 'overComposite1.k2.baseVal = .1')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k2' propoerty of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL Math.round(overComposite1.k2.baseVal * 1000) should be 900. Threw exception TypeError: 'undefined' is not an object (evaluating 'overComposite1.k2.baseVal')
